[Support][DebugInfo] dbg Value to Declare converter can reuse stack slot for identical conversion expressions (#50228)
Contributes to [MOTO-837: \[Debugging\] Values are sometimes incorrectly
marked as `variable not
[Internal link]

I know this function is probably going out the door at some point (and
is only used for CPU now), but this minor fix helps slightly with saving
stack space, and makes it easier for the backend to preserve these
values. With this fix, we can now see variable `p` throughout the loop
in the code in the ticket.

MODULAR_ORIG_COMMIT_REV_ID: 6a80c6053ad0ae85562b28d90136f695311b6803