[MO] Add a pass to validate data types (#50196)
In an least one case, we cannot support a data type on an architecture
(bfloat16 on ARM CPUs). This change adds a new pass which checks input
and result tensors to the graph to determine if the data types for the
tensors are valid for the given target architecture.

Related to GRA-1047

MAX_GRAPH_API_ORIG_REV_ID: aea886ddc3cee2d7386da96d43941711bd73a6a8