[vscode] Add telemetry wrapper (#50119)
Adds a wrapper around the Microsoft telemetry library we're using to
allow us to perform a few different functions:
- Add common context to the telemetry reporter (this is actually
  supported by the VS Code telemetry primitives, but isn't exposed to us
  via the library).
- Allow us to transparently enable/disable telemetry reporting.

Currently we have no common properties to report - we'll get there,
though. One thing we know we'd like to report is whether the user is a
Modular employee.

We disable telemetry reporting in development runs (where the extension
is launched from an IDE) and in test runs. In both cases, the telemetry
isn't something we're interested in. If you're running the extension
from VS Code as a debug session, you won't have a connection string to
send telemetry anyways, but for some CI situations we run with a
telemetry connection string available.

The connection string is now set as part of the extension build step to
prepare for a potential open-source release of the extension - we don't
want forks of the extension to report telemetry to us without deliberate
action by the fork author.

MODULAR_ORIG_COMMIT_REV_ID: d297e7963bd91aa1f53824bfcc621743551aa82a