[Kernels][layout] Add mask parameter to LayoutTensor (#50230)
We are back to the explicit masked tensor parameters approach but with a different way of handling the bounds.
In this PR, the runtime layout is clipped to determine the actual size.

In that particular case, the RuntimeLayout object may contain a fully static layout,
but the actual shape is different, which requires checking bounds.

KERNEL_BENCHMARKS_ORIG_REV_ID: f02206e7cd8d50d3424552f071f55e99b92c64e8