[Serving] Use OTEL prometheus exporter (#50407)
We are switching metrics to the OTEL SDK to enable exposing metrics
through the Prometheus API and Motel, while avoiding duplication of
instrumentation. This change also centralizes the definition of MAX
metrics to encapsulate the OTEL dependency from the points of
instrumentation (similar to the status quo for logging) and helps us
keep tabs on all the metrics we've instrumented.

One proviso is that multiprocess metrics are not supported out of the
box with OTEL's Prometheus collector, but currently all of our metrics
are instrumented in the API process. If we find that we must instrument
a metric in a worker process, we'll have to pass that data through IPC
to the API process.

SERVE_ORIG_REV_ID: 35e9440765a05fe5610e99ef66e881da761e2613