[KGEN] Add warp_size to the target attr (#50460)
The warp_size is needed for the target and we need to be able to query
it on both the CPU and GPU. This stores the optional warp_size in the
target attr and allows one to compile-time query it. It leaves filling
the warp size to the library.

Part of KERN-1136

MODULAR_ORIG_COMMIT_REV_ID: 52c4809536a74bfa0892da4a14ed6d846c4fa2fe