[MOGG][NFC] Move mogg_register->register_internal (#50503)
This is a wide an shallow change to rename `mogg_register.*` to
`register_internal.*`. These decorators are now part of our internal
API, not intended to be exposed to users. However, the presence of
`mogg_` in the name does leak some of the abstract out to the internal
users.

The name `register_internal` more clearly represents the use of this
decorator, for internal use when registering a function with the graph
compiler.

Fixes GRA-1199

NN_ORIG_REV_ID: 8b9ccaf52dc93da87e03499b6c6b1a8ace445e60