[stdlib] Format floats using dragonbox algorithm
Uses the dragonbox algorithm from C++ std::format to separate a float
into significand and exponent with roundtrip guarantee, and shortest
representation. The float formatting itself now matches Python
logic where < 0.0001 or positive exponent > 15 will use scientific form.

- [Paper](https://github.com/jk-jeon/dragonbox/blob/master/other_files/Dragonbox.pdf)
- [Reference Implementation](https://github.com/jk-jeon/dragonbox)

This is required for synching GPU printing of floats across threads. It
also greatly improves the representation of floats, for example:

- Float64(5.6) used to be represented as 5.5999999999999996 even though
this is the equivalent of 5.6 when represented in binary.

- Float32(5.6) used to be represented as 5.5999999046325684 due to
upcasting to Float64, but is now correctly represented as 5.6 when
printed or converted to a `String`.

ALGORITHM_ORIG_REV_ID: 969350a8b13bdeee0aa89757aa896ed21e021cfc