[Serve] Add timeout to create batch logic V2 (#50616)
This PR adds the timeout period when create a new batch. That logic
exists in the dynamic worker implementation

[Internal link]

but was missing in V2 intentionally as it wasn't the happy path for perf
related work. Now it's time to align the implementations.

SERVE_ORIG_REV_ID: 786d7c89ec96124a90fe1fe92bf6bdc9436f2b07