[MOGG][Extensibility] Support lambdas for variadic inputs. (#50053)
Add support for TensorSpec and lambda fusion for kernel with variadic inputs.

- Created a new compiler.specsof directive that returns a tuple of specs for variadic operands.
- Updated AutoSpecialize and Slicing pass to work with the tuple of specs.
- Updated MoToMOGG to parametrize correctly the tuple of specs and handle the variadic lambdas.

One big difference between NDBuffer / New API is that with the NDBuffer API all the fusion lambdas are created in MoToMOGG,
but with the new API those lambdas are created offline during kernel compilation (SliceMOGGDPSFuncs).
The problem with fused variadic operands is that we don't know during offline kernel compilation the number of lambdas we need.
To solve this, we first create a single unique lambda for the variadic, that load the first value of the variadic.
Then in MoToMOGG, we clone this lambda N times, and for each lambda we update the body to load the correct value of the variadic.

Needed for concat input fusion support.

MODULAR_ORIG_COMMIT_REV_ID: a7c933bb1d6cf32edf5011bf17f5afa51855fb03