[mojo-lang] Enable `__init__(out self)` syntax.
As discussed in [this public
[Internal link]
the use of `inout self` in initializers is incorrect. Initializers don't
actually read the
argument, they only write it. Per discussion on that thread, we're
moving to 'out' as
the keyword used in function declarations, which allows us to spell the
function
type correctly as well.

This keeps support for the old syntax for migration support, but we
should
eventually remove it.

MODULAR_ORIG_COMMIT_REV_ID: 115768e0c6df069f3cf2e4b3d806db1cc2e6d8a2