[mojo-stdlib][Kernels] Migrate to `__init__(out self)`
This adopts the recent changes that allow the use of the `out`
argument convention.  This argument convention more correctly
models the nature of `__init__` which initializes self but never
reads from it.

This was discussed on this public thread:
https://github.com/modularml/mojo/issues/3623

GPU_TEST_ORIG_REV_ID: d05be1ccb28b254aeccd5807858546aeb1777991