[Kernels][layout] propagate masked in vectorize (#50681)
Have been thinking more about masking element layout. I don't think it is possible in our case
becuase of the following reason. Suppose we have a tensor of 7x7 and we tile it by 4x4, the edge
tile will have shape 3x3 and static layout of 4x4, now if we vectorize it by 1x2, ideally, some
element will have layout 1x2 and edge elements will have layout 1x1. If we masked element layout
in vectorize, the entire tile will have element layout of 1x1 since vectorize does not understand
the coordinates of the element so it does not understand bounds.

With that said. In our case, since M and K dimension in matmul is fully vectorizable. There is no
need for us to mask element layout for now. This PR simply propagate masked in vectorize. If in the
future we need to handle situtions described above. We can provide a method called `unvectorize` which
revert the vectorized layout to scalar layout. CUTLASS also does the same thing although it will
automatically stash scalars if a layout is not vectoriable.

LAYOUT_ORIG_REV_ID: 048b54cd8c4020263fc6d755067a128ba7278250