[KGEN][AMD] Start plumbing the target accelerator into KGEN (#50719)
This is a first step in making Mojo and MAX aware of the target
accelerator and does not rely on some env flags or other hackery to
support other types of accelerators.

This starter PR first adds the accelerator arch into the target attr it
then uses the device context to populate the accelerator arch in the
mojo entry points.

Based on the ideas in
[Internal link]

Part of MOCO-897
Part of KERN-1161
Part of KERN-1136
Part of KERN-1170
Part of RUNP-544

MODULAR_ORIG_COMMIT_REV_ID: 2e92ffe2b47b5641309cb04427f9ea939b14e6ff