[Runtime] Remove CUDARuntimeOptions (#50732)
This code was unused at run time, so remove it. A number of tests did
specify the `--cuda-device-id` argument, but all of them used device 0,
and that value was not actually used at run time in any meaningful way.

The in-progress multi-GPU support handles specifying the devices in the
graph, not via a command line argument.

So remove this code to make the implementation of AMD support easier.

Completes GRA-1256

MODULAR_ORIG_COMMIT_REV_ID: 07fd4bec0e59857a11627adc501c24012138847d