[GraphCompiler][GenericML] Add  `*deviceIndices` extraAttr for mgpModelOp (#50849)
This adds a graph-compiler populated attribute that specifies the input
and output's devices. It populates indices which refers to a specific
deviceRef in the `devices` attribute (which is all the devices used by
the model).  This allows users of a MGPModel to determine which device
is used for each input/output.

This also allows MEF and users to restrospect into an MGPModel and
determine the devices of the input/output without turning on the
Metadata section of MEF.

Towards AIPIPE-115

MODULAR_ORIG_COMMIT_REV_ID: 70c355bd9c3188d562ac0506658401833393a878