[Kernels][GPU][AMD] Fix test_prefetch.mojo for AMD GPU. (#51241)
Test did not specify target so even on AMD GPU targets it tested for the
nvidia output (but compiled with AMD GPU implicitly).
Now it specifically compiles and tests for nvidia and amd gpu (in case
of AMD there is no corresponding prefetch to `prefetch.global.L2`, so we
default to `llvm.prefetch` which should be a no-op for AMD GPU).

GPU_TEST_ORIG_REV_ID: 4f5f5a240e7144417b30319499999f8069a65f23