[MGP] Fix debug_assert for mpg.tensor_spec.get_dim (#51339)
The len of a StaticTensorSpec is the total number of elements rather
than the rank. When running in debug mode, this caused a lot of
unexpected failures, as the assertion was intended as a constraint on
the rank of the StaticTensorSpec.

Change the debug_assert to a constraint so that it is always checked.

Toward: GEX-1421
MOGG_ORIG_REV_ID: 0a7181a2ca79bd4a41178b6957e096f23766a72b