[MGP] Add `mgp.buffer.get_size` op (#51366)
This is used to determine the runtime size of a buffer. This is used by future commits in this stack to determine if the memory pool used by dynamic memory planning needs to be resized or if it's big enough.

Towards GEX-866

MOGG_ORIG_REV_ID: 588c3a1c8ca3e660e45835d5e6d7a94a04f6a7ec