[graph-api] Reland: Switch to an opaque location (#51421)
[graph-api] Reland: Switch to an opaque location

This opaque location is a new RMO attribute and enables using the fast
path during graph build time.
~~Getting the current stack frame in python is essentially free, while
building a traceback is quite slow.~~
Extracting the stack is relatively fast, but not quite as fast as just
grabbing the stack frame.
That said, it avoids references cycles in python. So we extract the
stack.
To later get the stack trace, at the bottom of the stack we call into
python.
Python then generates a string traceback.
This has a second benefit of only generating the traces for asserts that
are in the final graph. Anything pruned by shape inference has no cost.

To avoid major linking headaches with python, a function pointer to
generate the stack trace is passed in via the opaque location.

This leads to llama3 graph build time dropping from ~12s to ~3s (would
be ~2s with current frame, but with reference cycle problems).

-----

Now with a fix to only release the gil for max graphs.
For pytorch, the gil is required for constant folding.

MAX_PYTHON_ORIG_REV_ID: 88bdb94df07725580dd01553536d594855533672