[KGEN] Add time trace to bytecode interpreter (#51476)
Start instrumenting function evaluation in the interpreter. This allows
understanding what functions are hit the most in the interpreter so that
kernel devs may tune their comp-time code & compiler devs may optimize
the interpreter.

For now this is implemented with our builtin TimeProfiler, which means
it'll only work when mojo is built with tracing enabled. This is OK for
internal usage, but at some point we may want to decouple it so that
mojo comp-time profiling is not dependent on profiling the compiler
itself.

I've adjusted the `kCompiler` tracing type so that existing levels are
incremented, and level 1 now only profiles the interpreter. Happy to
adjust this to an independent tracing type if that works better.

Example usage:
```
bb KGEN/tools/kgen --//:profiling=interpreter        # this corresponds to 0100000
kgen -elaborate -time-trace -o /dev/null <mojo file>
```
Then load the trace into perfetto or speedscope.

MODULAR_ORIG_COMMIT_REV_ID: 21d1ef3916f430e72b79841f4e3403d2136b877f