[mojo-stdlib][KGEN] Expand constrained to work with a `String` parameter
This introduces a new `kgen.param.assert.ex` operation that takes a
pointer+length for data as a parameter, and uses that to introduce an
overload of `constrained` that takes a general `String` instead of just
a `StringLiteral`.  This allows more interesting expressions, but at the
cost of compile time.

The existing overload of `constrained` is retained so the majority of
callers that pass in a string literal directly will not see any compile
time impact.

MODULAR_ORIG_COMMIT_REV_ID: 724d791274d08a582b6c738d445fdac90e9fef72