[GenericML][NFC] Rename cuda to cpu in TableGen docs (#51529)
We have a number of examples of device spec usage in TableGen
documentation which use "cuda" to mean "gpu". With the changes in:

[Internal link]

landing, we also should update this documentation to reflect the proper
way so specify a GPU device in a device spec.

Towards GEX-1269

MODULAR_ORIG_COMMIT_REV_ID: 9ee4011ef2c5c17f104b99469210131343c525f2