[KGEN] Set the low-bits for Nan to 1 (#51778)
A NaN for F8E4M3 is of the form

NaNs: S.1111.{001, 010, 011, 100, 101, 110, 111}

and similarly for F8E5M2 is of the form

NaNs: S.11111.{01, 10, 11}

before we were always setting the lowbits to be 0, this changes the
generation to set the low-bits to 1. This makes the floating point
representation aligned with what is generated by CUDA libraries.

Closes KERN-1260
MODULAR_ORIG_COMMIT_REV_ID: 72fa7e5a781e058a7be238fef3038577ec370c56