[SDK] Fix type of max.profiler.tracing.traced (#52040)
The type was imprecise before, and basically destroyed the type
information of anything that was annotated with it. This turns it into
two overloads that use generics when appropriate. To test this out,
consider this code (you can test this by putting it at the end of
`tracing.py`):

```python
def example() -> None:
    @traced
    def mything(a: int, b: str) -> str:
        return str(a) + b

    @traced(color="red")
    def myotherthing(a: int, b: str) -> int:
        return a + int(b)

    x: str = mything("2", "a")
    y: int = myotherthing(1, 5)
    print(x, y)
```

This is wrong -- `mything`'s `a` is being passed a `str` when it should
be an `int`, and `myotherthing`'s `b` is being passed an `int` when it
should be a `str`. Before this change, MyPy is fine with this, since
`traced` destroys the type information. After this change, MyPy is
correctly able to diagnose the error, and upon fixing the type issues,
once again succeeds.

MAX_PYTHON_ORIG_REV_ID: a15a95b9bad8cc945e91fd98d1206400ad05e672