Revert "[Pytorch] Move torch/max dtype mapping to common utilities (#52001)" (#52147)
This reverts commit 1604f7f27992b539fd8a1917960f355c6015d263.
Adding a dependency on `modular.utils.misc` broke the serving benchmark

[Internal link]
Also we should probably rethink the name of this since it is exposed
externally (in other words it shouldn't be in the `modular` namespace
package).
So prefer to revert for now and we can reland it with a packaging test.

MODULAR_ORIG_COMMIT_REV_ID: d4c6096f360e62573a890f129f64fd374b983e42