[stdlib] cleanup: Remove `_lit_origin_union`; use `__origin_of(a, b)` instead
This is a low-level operation that was only used in one place in
practice. With the recent
change to support using `__origin_of(a, b)` as a way to form origin
unions, `_lit_origin_union`
is no longer strictly necessary, so let's remove it.

MODULAR_ORIG_COMMIT_REV_ID: 3128b8776cea534b0b58525368f2865805b57237