[SDK] Rename cuda_device_count to accelerator_count (#52154)
This change renames a `max.driver` API to remove the word "cuda". This
API is not specific to NVIDIA GPUs, but instead should return the number
of GPUs of any type available to the host. Based on the design
discussion in GEX-1445, we will use the word
"accelerator", which is more general than "gpu". This is a continuation
of [Internal link]

Towards GEX-1445

MAX_PYTHON_ORIG_REV_ID: 931d1af24dc64aec1ce5e7d070b7a85bf8e095cf