[Serve] Fix CUDA initialization issue (#52210)
This PR fixes
[Internal link]
The problem is that [the code that gets the GPU memory and other
[Internal link]
initializes the CUDA driver and it happens in the API process. MAX Serve
is a multi-process application and loads models in the Engine process
which also initializes the CUDA driver. CUDA doesn't like it and crashes
with `CUDA_ERROR_NOT_INITIALIZED` in [our CUDADriver
[Internal link]
while calling CUDA's
[cuInit()](https://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__INITIALIZE.html).
This PR just removes the "system.gpu" telemetry entry. We should find
another safer way to get the data.

---------

Co-authored-by: Keith Smiley <keithbsmiley@gmail.com>
Co-authored-by: Patrick Dougherty <patrick.dougherty@modular.com>
SERVE_ORIG_REV_ID: 71b1b1784817411913d8b50382b6bfef8ed2596c