[Kernels][Pipelines] increase our max sequence length in MHA kernel to 16384 (#52268)
We hard coded the max seq length to 16384 in our FA kernel after adding
ragged tensors, because this moved all length metadata to the GPU. We'll
need to pass around a copy of max_seq_length and max_cache_length on CPU
in order to set this argument to the correct amount, but in the
meanwhile we're bumping up the arbitrary limit and adding validation in
the pipeline. I benchmarked this change and noticed the exact same RPS
before and after.

Slack thread
[Internal link]

Long term fix tracked by KERN-1104

NN_ORIG_REV_ID: 89e45a337793e57a63e7298708efe91cf38f4aeb