[Driver][SDK] Launch model on the same thread (#52081)
We have multiple places that launch the model asynchronously and then
just block waiting for the results anyway. This incurs some quite
wasteful overhead. We can just run on the same thread. The user has the
power to launch this in a separate process or thread if they so wish, we
don't need to fork

Without this it takes 80-100us to reach the first primitive from the C++
entry point (executeDeviceTensors) With this it is down to ~20-40us.

Only one part of the stack relied on the model launch being in a new
thread and that is some heavyweight asserts behind MODULAR_PARANOID. We
aren't really using that so we can just remove that assumption.

MODULAR_ORIG_COMMIT_REV_ID: 0d0ec28cff767da3525421d17994b026ecf8ff91