[mojo-lang] Fix bugs related to SRValue -> ref conversion.
This fixes [Internal link] a collection
of issues handling emission of SRValues to argument conventions that
require a memory argument.  In these cases, we need to emit the value
to memory and re-infer.  This was completely broken for indirect calls,
and didn't handle the case of direct calls that failed to re-infer.

MODULAR_ORIG_COMMIT_REV_ID: 172c9efe1ca9083e50f1f58629d0b990bb3090cd