[External] [stdlib] Deque - don't use non-self out args (#52550)
[External] [stdlib] Deque - don't use non-self out args

As the original author of the Deque, I would like to propose to not use
the `out` args instead of `-> ElementType as element` in `pop()` and
`popleft()` methods.

The named return variable was just a simple convenience in these
functions and not needed at all.

In general, having something in the args list of the function that you
do not pass to the function at the call site looks weird (except for
constructors). In my opinion, it makes the function signature just less
clear to read and the return type of the function unnecessarily hidden
inside the args list.

So, let's just use normal return statements here and have a clear
function signatures.

Co-authored-by: Alvydas Vitkauskas <alvydas.vitkauskas@gmail.com>
Closes modularml/mojo#3851
MODULAR_ORIG_COMMIT_REV_ID: acae8a18cea7f91639d059e03dff88e3e3d1f5a5