[NFC][Mojo tooling] Rename DeclViews into PublicDecls (#52582)
This is merely a cosmetic change. DeclView is the wrong name for this
API. Instead, this API is used as a public interface around internal
decls for the purpose of tooling. Therefore, it's better to have a more
explicit name for clarity.

MODULAR_ORIG_COMMIT_REV_ID: e196777b61ee02a153782fff40dafb602c12b55a