[KGEN] SignatureGeneratorType helper (NFC) (#52121)
Add a helper type: `SignatureGeneratorType`, which is just any
`GeneratorType` that wraps a `NewSignatureType`. This is a direct
replacement for the existing `SignatureType`.

This change by itself is NFC. It decreases the diff for the [next
[Internal link] which will replace
`SignatureType` in `GeneratorOp` with `SignatureGeneratorType`.

MODULAR_ORIG_COMMIT_REV_ID: 1d9ae39cffc6701cdd25179f6e200f78fc5da022