[SDK] remove ContinuousKVCache from mojo pipelines (#52684)
The Mojo pipelines are on the path for deletion. I'm removing the
ContinuousKVCache wrapper now as it is only used by this integration
test and is in the way of me landing #52682. The base kernel type has
changed and there is no value in fixing this test case.

MOJO_MAX_ORIG_REV_ID: ab6bcb835851496a1b9a38adb0bf363f7e051661