[Kernels] Infer num_q_heads from q tensor not kv_params (#52693)
There was a slight mismatch between the num_heads in
KVCacheStaticParams, which is the number of kv heads, and the value
needed to calculate the alibi score bias, which is number of attention
heads. As such, we now grab the value for alibi from the dimensions of
the q tensor directly. This assumes that the dimension for the q tensor
will remain static over time.

NN_ORIG_REV_ID: 9ab69cfdc1a06a7e7d2f20487414b0e9de483c71