[KGEN] FuncInterface use SignatureGeneratorType (NFC) (#52337)
Adopt `SignatureGeneratorType` for `FuncInterface`.

Instead of a `getSignature` function returning a `SignatureType`, we
replace it with `getSignatureGenerator`, which returns a
`SignatureGeneratorType`.

All KGEN ops have already been migrated to use `SignatureGeneratorType`
already. The only remaining non-compliant op is `LIT::FuncOp`, which
will be migrated later. Currently it uses a forward-compat helper to
comply with the interface requirement, which is OK because LIT::FuncOps
are short-lived in the compiler & the only place that use `getSignature`
on them (if i didn't miss any) is in SymbolConstantAttr's symbol use
verifier.

MODULAR_ORIG_COMMIT_REV_ID: 244be97b3758b63fce1449806a0df1158dab1752