[MOGG][Extensibility] Add new function to build a partially static IndexList (#52289)
Define a new function that takes a compile-time DimList, and a runtime IndexList,
and create a runtime IndexList that get the values from the dimlist when available, or the runtime list otherwise.
This helps a lot with LLVM optimization when compiling kernels with partially static shape.

Eg: if you have a shape like `[3, din0, 7, din1]` the compiler will see and optimize through the static dimensions 3 and 7.
BUFFER_ORIG_REV_ID: 935154152563aaf7e40eb6ca6096e9c822a9a75b