[stdlib] Change `Dict.hash` data member type to be `UInt64`
The `Hashable` trait has `__hash__` function return a `UInt` hash value.
So, `Dict` storing the hash value as an `Int` is incorrect after the
change in the return type from `__hash__` when it used to be `Int`.

In the near future, `__hash__` will be changed to return a `UInt64`
(whcih also aligns with the new data-flow hashing in progress). So, go
ahead and update the `Dict.hash` data member to prepare for this change
(this is why we went straight to `UInt64` instead of `UInt`).

MODULAR_ORIG_COMMIT_REV_ID: f6782bb4972e7b1c8ee2fc43fc7120105065e426