[KVCache] Refactor KVCache managers to accept prompts in 'fetch' and 'step' (#53151)
This will enable prefix caching in later PR.

In `fetch`, the kv_manager will look at the prompt and attempt to match it on the radix trie. If any tokens in prompt are matched, then we will add those blocks to the sequence's cache so they do not need to be redundantly computed.

In `step`, the kv_manager will add the newly commited blocks corresponding to the tokens in the prompt into the trie. This will allow blocks to be reused in future invocations of `fetch`.

Towards AIPIPE-217

MAX_INTEGRATION_TESTS_REV_ID: c2da3da0b915e96261cd8949bdc7d528c8a812e3