[mojo-lang] Replace `InitSelf` convention with `ByRefResult`
This replaces the use of the InitSelf convention with the existing
ByRefResult convention, making inits static functions that return
Self.  This simplifies the compiler by removing a concept entirely.

As a consequence of this, the 'self' argument to inits is passed as
the last argument in the IR instead of the first.

MOGG_ORIG_REV_ID: 8d6124b11312eadace129cc1bc3344d4c13915ba