[mojo-lang] Form variadic packs with correct variadic list element type (#53366)
When forming a `VariadicPack` due to type list unpack syntax, the
list of types may have arbitrary trait type, so the metatype of the
variadic list could be anything.  That said, the trait specifier
parameter for VariadicPack upcasts to an `AnyTrait[AnyType]` metatype
and the variadic type list needs to match.  We handled the trait
type upcast correctly but not the variadic type list, causing us to
reject the testcase.

MODULAR_ORIG_COMMIT_REV_ID: 395d604da84469f30ad3596ee7344777a38a6246