[mojo-lang] Simplify handling of parametric metatypes. (#53370)
This completely removes the code for handling "zero cost" conversions of
values with parametric type using a rebind.  We handle these correctly
with the existing `emitMetaTypeToTraitConversion` codepath, and it has
the advantage of forming a TypeConstantAttr for a metatype bind, rather
than using a rebind. The problem with rebind is that it propagates the
wrong type through ref-returning functions, e.g. `get_element` in the
ownership_refs.mojo test, and `VariadicPack.getitem`.

This temporarily breaks a testcase I recently added, but I will address
that
later.

MODULAR_ORIG_COMMIT_REV_ID: 44efc681745b3b91717281c1d29760a27b3d1728