[LSP] Ignore multiline semantic tokens (#53506)
The LSP protocol does not allow semantic tokens to span multiple lines.
At some point in December, we started incidentally creating a symbol
with a range that spans multiple lines, breaking this invariant and
causing us to accidentally return horribly invalid data due to a signed
-> unsigned conversion. This manifested as garbled highlighting when the
LSP was active.

To fix this, I've made the language server ignore symbols that span
multiple lines for semantic token purposes. This is a bit of a
heavy-handed solution, but any other fix is a bit difficult at this
point because we don't know how long to make each span. If we need to
report this kind of data in the future we can revisit this, but for now
this maintains the status quo without garbling highlighting.

Fixes MOTO-903.

MODULAR_ORIG_COMMIT_REV_ID: 2989704e8db509d29b7c28c3fb55d6a48226a32f