[KGEN] Rename `KGEN::ParamRefType` -> `KGEN::ParamType`, NFC.
This renames a (very old) core kgen type to better align with that
it does: it projects a TypedAttr into a Type.  It is itself not a
"reference" to a parameter decl like `ParamDeclRefAttr` so using
that word is confusing, this is also more concise and consistent.

MODULAR_ORIG_COMMIT_REV_ID: fa2b6ffbb099b8fcdab5ddd83d5990f348e8c7b0