[mojo-lang] Introduce `#lit.upcast` for trait upcasts. (#53750)
This attribute is used to convert a type expression from a derived
trait to a base trait.  In the case when the input is a type (like
int) it constant folds to TypeParamAttr, but is needed in casts
where we need to propagate an upcast to elaboration time.

This is NFC for the compiler as of today because nothing can
correctly exercise this path, but matters for #53374 which needs
to propagate typeValue in this case, which TypeParamAttr doesn't
do.

MODULAR_ORIG_COMMIT_REV_ID: db4ff171cffe14290d9b716e85c01cc13dd0ff13