[KGEN] Use xxhash to hash the function name instead of blake (#53996)
We do not need cryptographic hashing for the name of the function, so
just use xxhash. This is prep work to remove StringLiteral.hash as a pop
op which is required for the split compilation work.

MODULAR_ORIG_COMMIT_REV_ID: 4470bdbef8a41287bc56e713a13b0da3c551c3a7