[MOGG] Remove loc from call context (#54082)
The location is used when reporting errors from the kernel, however it
has two problems.

1. The location is the location in the MEF, which is useless to user
2. Even if it was the graph location, it's still a fused graph location
which isn't going to be much use to the user since that includes all the
ops.

So we should remove it rather than reporting a location to the user
which doesn't make any sense to them. We could resolve this by encoding
the location into the assert but then this has the problem of blowing up
compile time

MODULAR_ORIG_COMMIT_REV_ID: ac295216921c4886fb316749a80958dc8e292c14