[KGEN][DebugInfo] Restore recursive DIStruct emission to fix crash (#54071)
Fixes [MOCO-1524: \[mojo-lang\] Compiler crash in debug info on
recursive
[Internal link]
Fixes [MOCO-1544: Crash in
[Internal link]

For some background, the original DIStruct emission used a basic
AttrTypeReplacer, and detected recursion by checking whether a field
type was successfully replaced or not. If not, it must have hit a repeat
in a replacer cycle. So the emission logic simply truncated the nesting
by returning a none type.

[Internal link] updated the replacer
infra to use a CyclicAttrTypeReplacer (due to being needed in other
places), which is more powerful in that it allows explicitly specifying
the cycle-breaking behavior. However, it is also powerful in providing a
cycle-aware cache. This had the side effect that more than two instances
of the same type might end up being emitted (e.g. one inside a cycle,
one outside) depending on which node was used to enter the cycle. Having
two types with the same name in dwarf confused the dwarf parser, and
caused the error today.

This PR restores the original behavior by parameterizing
`DebugInfoTypeConverter` over whether the replacer is cycle-aware or
not. This allows the LowerLITTypes code to still operate by truncating
eagerly, while preserving the new behavior during kgen-to-llvm lowering.
This parameterized version won't be needed once we fully convert over to
the new flow.

MODULAR_ORIG_COMMIT_REV_ID: 9d66f69d11f1e0aa1b4869ec6b5be15eff1b4adc