[Kernels] Make `largest` a parameter of `topk` (#54138)
Callers know when they want top k or bottom k at compile time, so make
it a parameter rather than a runtime value.  This was noticed as part of
the differences betweeen the CPU and GPU implementations of topk.

Adjusting it to be a parameter will help with unifying the
implementations for topk.

GPU_TEST_ORIG_REV_ID: f6f0d8663a455a4b52c7e63275a647258ba271b2