[KGEN] Dumping ASM Will Produce A Unique Name (#54102)
Stacked PRs:
 * __->__#54102

--- --- ---

### [KGEN] Dumping ASM Will Produce A Unique Name

For the object compiler, the current behavior dumping files is if we
produce multiple artifacts, we write to the saveTempsPrefix but
overwrite everytime we dump. This is unintuitive and may make it seem
like the compiler only produced a single artifact -- where the user is
really just seeing the last compiled artifact.

This solves this by hashing the contents of the ASM before dumping. This
is on a slow debug path and xxh is fast (~30GB/s in one setup
https://xxhash.com/) so this should not be a problem.

MODULAR_ORIG_COMMIT_REV_ID: 9b61e1b830094cb5823e41d30324ec474ac8041b