[mojo-lang] Fix a lazy-emission of def-arg box mutations.
`def` arguments can be mutated, and therefore need a local copy of
the value, but we defer the generation of the copy until (and if)
the argument is ever mutated.  In the case when the argument is
accessed immutably and then later mutated, we need to make sure to
update the earlier accesses to refer to the local copy, because
loops and other control flow can cause it to be reached dynamically
after the mutation.

This fixes https://github.com/modular/mojo/issues/3955

MODULAR_ORIG_COMMIT_REV_ID: d53c484a54f6e82a9980da836bad5b2c46687215