[MOGG] Touch Up Or Disable OLD API Tests (#54175)
We still have tests depending on the old api. For example, we might be ingesting secondary test libraries with the old interface.

We fixup or disable tests with issues to make them work without the old api. In places where we are testing the same concepts in both the old api and new api, we just delete the old api tests.

MOGG_ORIG_REV_ID: da771efebc43258f4a0bf292382d0bdd34717443