[mojo-lang][CheckLifetimes] Refactor `scheduleNeededDtors`, NFC (#54254)
This makes the control flow more clear, e.g. what happens when in
dryMode and what the behavior is.  This is still buggy and needs to
be fixed, but this is a logical step to clarify things.

MODULAR_ORIG_COMMIT_REV_ID: dcdd7e48ad0c28f1255cde5ed07569e82758ea0e