[LIT] LIT::StructType do not carry metatype (NFC) (#54028)
Avoid carrying the metatype in LIT::StructType. Removing it is the first
step towards turning `AnyStructType` into `MetaType<StructType>`.

Unfortunately, things are not always so straightforward. Several places
are looking inside the metatype field. Casting it to `AnyStructType` and
getting the `TypeSignatureType` out in order to figure out the generator
signature of the struct. Theoretically, cases like this should not be
using a `StructType`, but a `ParamType<BindParam<SymbolRefParamAttr>>`.
But to make things more practical, we can just have `StructType` carry a
typed `SymbolRefParamAttr`, and perform the dual duty of binding params
& converting type-value into type. This will be cleaned up as we remove
`TypeSignatureType`, but for the current PR, we move incrementally, and
simply stuff the `TypeSignatureType` into `StructType` to minimize the
diff.

This `TypeSignatureType` will go away in the next stack of PRs where we
update `StructType` to carry a typed reference.

MODULAR_ORIG_COMMIT_REV_ID: 883250c02e522a2ab4755a5729ea182b24c4808b