[LIT] Simplify LIT::AnyStructType (NFC) (#54132)
This is a preparation step for the next PR.

For this PR, we make it so that `LIT::AnyStructType` only stores a
`LIT::StructType` & update tests.
In the next PR, we simply replace all usages of `LIT::AnyStructType`
with `LIT::MetaType`.

Note that the double angled brackets in the IR will make sense in the
next PR where we no longer always know it's a struct type inside.

MODULAR_ORIG_COMMIT_REV_ID: c21b621ebe4e2fc45d538a132609385c4fe39a1c