[KGEN] NewSignatureType rename (#54284)
Rename NewSignatureType to FuncType & FnType for KGEN & LIT respectively
(`!kgen.func`, `!lit.fn`).

Also rename SignatureGeneratorType to FuncTypeGeneratorType &
FnTypeGeneratorType, and related methods like
- `getSignatureGenerator` -> `getFuncTypeGenerator`.
- `remapToSignatureGenerator` -> `remapToFuncTypeGenerator`.

---

Rationale:
The word "signature" by itself indicates that it's a type. This is nice
because when you say "signature", you know it's a type. But the word
"func" by itself describes a function value, so in order to distinguish,
we need to say "FuncType" instead of just "Func". This means the
replacement of "SignatureType" is "FuncType", but the replacement of
"Signature" is also "FuncType".

MODULAR_ORIG_COMMIT_REV_ID: 9fec3a9d4e38cf69bc7c3215df1784a33233af31