[SDK] Simplify the is_float in DType to not know about float dtypes (#54219)
The problem with the current approach is that one has to enumerate all
the float dtypes which becomes a bit anoying as we add new dtypes. This
just uses the masks (which is also efficient) to check if the value is
floating or not.

MAX_PYTHON_ORIG_REV_ID: bd94b109c582ee3f1a6ac6bfe0e7ede4abf74e52