[mojo-lang][CheckLifetimes] Fix a bug in dtor emission.
After all the refactoring and semantic change to __disable_del
this Fixes MOCO-1558.

The basic issue is that we need to accurately find the right level
of object to destroy when seeing the last use of it, even when a
subobject has been replaced or disable del has been used. For
example given:

```
   init(&aggregate)
   use(&aggregate.field1.subfield)  <<-- We are here.

   init(&aggregate.field1)
   __del__(&aggregate)
```

We need to know that we have to destroy `aggregate.field1`, not
aggregate or `aggregate.field1.subfield`.  This corrects the
previous logic to be general, instead of just handling the full
object case.

MOJO_MAX_ORIG_REV_ID: d35f040dec693d0347d6fee92af7a7179a9c9488