Revert "[KGEN] Move litcfg to BUILD.bazel (#54518)" (#54676)
This not having a litcfg makes copy pasting failed tests a bit harder
because of `-mojo-search-paths`. In the litcfg this file is absolutized,
in the BUILD file version this info is communicated via an env var
instead, which makes it harder to copy paste. Putting the path in the
BUILD file doesn't work as easily because there are tests with varying
levels of distance for searching up for the path, ex `../test-packages`
vs `../../test-packages`. This litcfg isn't the only one left so it's
fine to leave for now.

This reverts commit a7c9fd87d5bc828d68f518a1925112731f69a257.

MODULAR_ORIG_COMMIT_REV_ID: 9ee20b560505f6fe5dda02da21fc9e73c546c729