Remove the Examples tab from the Mojo documentation site
We're retiring the Mojo notebook examples in the GitHub repo, which is
the source for the current Examples tab on the Mojo documentation site.
We'll have a new interface for exposing tutorials and examples on the
site in the future.

MODULAR_ORIG_COMMIT_REV_ID: 385ba3732165d8cf9b0de6208d8e6cbaa63b314b