[Kernels][GPU] Switch the default compilation for NV GPU to be CUBIN (#54836)
This switches the code so that we compile to a CUBIN rather than PTX as
the default. This should lift the overhead of JIT to be entirely at JIT
(graph compilation time) rather than be partially at graph compile time
and partially at runtime (model execution time).

I've looked at the perf and it's marginal improvement/regression, but
this is a good thing to do in general. I'm also going to revisit the
todos in a followup PR and create the relevant linear issues.

closes KERN-1231
closes MOCO-1576
closes MOCO-1054
closes MOCO-1482

GPU_ORIG_REV_ID: cb8696570e4c18e0f4019d72d109527a8ba3a1d5