[KGEN] Add `kgen-debug-only` option to dump internal information (#54711)
[KGEN] Add `kgen-debug-only` option to dump internal information

The option is similar to LLVM's `-debug-only` with the exception it
allows to specify "verbosity" level. That option may allow to debug
complex passes such as Elaborator without a need to add custom prints
every time.

The way specified level in the option works is following: if level
is greater than level specified in `KGEN_DEBUG`, the code in
`KGEN_DEBUG` will be executed. For example
```

KGEN_DEBUG("stack-reuse", 5, llvm_unreachable("gotcha"));

% kgen-opt -kgen-debug-only=stack-reuse:5
```
the `llvm_unreachable` will be executed. However if user runs
```
% kgen-opt -kgen-debug-only=stack-reuse:4
```
the `llvm_unreachable` won't be executed.

Similary to `--mlir-print-ir-*`, in order for that option to function
correctly, multithreading needs to be disabled

MODULAR_ORIG_COMMIT_REV_ID: 161d65aa32a10eb933e2f3aaac73416581428800