[Serving] Add threading support to ProcessControl (#54984)
Add threading support to ProcessControl.

The the difference in the threading variant is that threading using a
normal c_int64 for heartbeats rather than a synchronized one. So long as
there is a GIL, this is safe. So long as writes are atomic, this should
also be safe.

Closes MAXSERV-868

---------

Co-authored-by: Matt Terry <mrterry@modualr.com>
MAX_TESTS_TESTS_ORIG_REV_ID: 3f7ce8d7de7568037e6fb6104f432d6a507305f1