[Kernels][GPU] Make the launch bounds parameteric on accelerator (#55134)
Stacked PRs:
 * __->__#55134
 * #55133

--- --- ---

### [Kernels][GPU] Make the launch bounds parameteric on accelerator

This adds a new MAX_THREADS_PER_BLOCK_METADATA that would allow us to
specify the max threads per block for both AMD and NVIDIA GPUs in an
agnostic fashion. This was an optimization that was missing for AMD
hardware because of the lack of generalization.

Authored-by: Abdul <adakkak@modular.com>
NN_ORIG_REV_ID: 27427e34c266e7f69cbdfaad066da98950b80a96