[Support] Zero-initialize the out arg for fills (#55212)
The `M::getScalarOne` and `M::getScalarNegativeOne` functions assume the
output argument is zero-initialized. They will set only the bits
required to fill the data type of a given size.

A new assertion in an LLVM integration in the `APInt` constructor was
triggering for these tests, because we were filling the lower 16 bits of
a output argument, but the upper 16 bits might have some non-zero
values, meaning the final value passed to the `APInt` constructor was
too large to represent in bits.

For example, we expected the bf16 value to be: 0x000003F80 when it was
0x10003F80 in my tests.

Fixes GEX-1794

MODULAR_ORIG_COMMIT_REV_ID: 25b01a6a13e1458c9eee27a1cfff32ff812303ff