[LIT] LowerLIT cannot remove parametrically singleton struct types (#54779)
Fixes [MOCO-1557: Crash in KGEN::SymbolConstantAttr::get and weird error
[Internal link]

When removing singleton struct types during LowerLIT, cannot just
blindly remove all param values that have singleton type, since the
actual type expected by the generator may be a parametric type, and thus
cannot be removed from the decl for users not passing in singleton
struct types.

This PR adds a map from generator symbol name to a bit mask representing
the param decls that were dropped (always singleton types). When
updating `SymbolConstantAttr`, instead of checking the type of the
binding params (which reflect the "rebound" type), just use the bit mask
to drop params.

For BindParams, this is simpler because we know the pre-bound type of
the generator. So we can remove params based on the declared type
instead of the actual type (and in the future we may consider moving
SymbolConstantAttr to a similar typed-reference approach).

MODULAR_ORIG_COMMIT_REV_ID: 84d813ce67d1b25c4d7ff87ca356f59d96854455