[Serve] Add stop sequence detection to chat endpoint (#55635)
This PR adds stop sequences to the chat endpoint.

Expected behavior is that for
[non-streaming](https://help.openai.com/en/articles/5072263-how-do-i-use-stop-sequences-in-the-openai-api),
the continuation up until and not including the matching stop sequence
is returned. In the streaming case, no such guarantees are given - it
appears to be best effort although I did not experiment heavily with the
OpenAI API.

SERVE_ORIG_REV_ID: 628a4c1020c8dc3aa59433e9328d51060f8045d3