[CI] Specify required memory for large GPU tests (#55521)
These tests load models that fill a large percentage of available memory on an A10. If they are run at the same time on the same GPU, we can see OOMs. This PR updates the tests to claim the majority of the memory on the GPU so that nothing else can be scheduled simultaneously to avoid resource contention.

MAX_INTEGRATION_TESTS_REV_ID: 719688ea53bdbe25f1b708173f235e8603fd65a0