[AsyncRT] Align the array allocation for work queue threads (#55751)
The `WorkQueueThread` type  has a minimum alignment of 64 (since it
includes a `unint64_t` member. Ensure that the allocation of the array
of these objects is also aligned at 64 bits (or whatever the alignment
of the `WorkQueueThread` type happens to be in the future.

This PR also enable the alignment check in undefined behavior sanitizer.

Fixes GEX-1823

MODULAR_ORIG_COMMIT_REV_ID: 6ef8ee8665a8c9b681693fb332cc07817917c5da