[KGEN][UB] Align payload offset to Payload's alignment (#55925)
UBSan identified few places where interpreter tries to access 8-byte
aligned Payload via pointer with a 4-byte alignment. That's due to the
fact that `payloadOffset` is a result of a total allocation size of a
`BOOperation`, which is not necessary a multiple of 8.

The fix padds total size to multiple of 8.

MODULAR_ORIG_COMMIT_REV_ID: 2f3fdcfadff00a4bdc79e8a90af45bd0b0aff13d