[mojo-lang] Remove `refineResultValue`, a major use of the interpreter in the parser
Now that we have `@always_inline("builtin")`, we can remove these
old hacks that was intended to support dependent types better. This
will probably help compile time and simplifies things because we're not
calling into the comptime interpreter in the parser in this place
anymore.

This also updates `Bool` so we constant fold Bool in cond expressions,
which is needed for this.

MODULAR_ORIG_COMMIT_REV_ID: fad37c77c069106834608a0e7e07b0cbbfdcc452