[mojo-lang][KGEN] Add infra for nicer elaborator errors on attr folding. (#56076)
The Elaborator is responsible for specializing code: this involves
substituting concrete param values into function bodies, with the goal
of getting simple constants out the back.  It does a reasonably good
job of this, except with error messages.  One problem is that an
attribute can refuse to fold (because of some semantic problem) but the
attribute has no way to report that in a reasonable way.  The elaborator
special cases ParamOperatorAttr but otherwise gives a hopeless generic
error.

This patch implements the infrastructure to provide this functionality
as a new optional hook on the ParameterAttr trait, moving the special
case over to that hook.  This also makes sure to diagnose the RIGHT
attribute, which may be buried inside an expression tree, not at the
root being evaluated.

MODULAR_ORIG_COMMIT_REV_ID: 60b41333b4213076031f8b7835cfe92f730a6be9