[Kernels][stdlib] Reland "Improve numerical accuracy for cumsum on cpu" (and gpu) (#56217)
Reland #56109 which broke the PyTorch tests for bfloat16 on GPU.
The reason is that torch has accuracy issues with bfloat16 on GPU, which
only show up now that MAX has increased accuracy.
The solution was to disable the GPU part of the PyTorch test, and also
add a standalone GPU test for cumsum

As an example:`cumsum(tensor(500).fill(1.1f))[-1]` gives

- `552` for max
- `532` for torch
- `550` exact formula (`500` is not representable in bfloat16, but `548`
and `552` are the closest)

## Original description:
This PR:
- Increases the numerical accuracy for cumsum operation
- Earlier, cumsum of a 1024 long vector filled with 1.1f would diverge
against PyTorch.
- This is now fixed, tests are added both in mojo and python

NN_ORIG_REV_ID: de4ad93be4fdc39e24390f60b0055b4d4c98d7b5