[Kernels] Remove deprecation from `DeviceContext.enqueue_function`
When enqueuing a function directly without compiling it first, it
consistently incurs 50-500 nanoseconds of overhead per enqueue when
reusing the exact same function and parameters. So this PR removes the
deprecation warning for when this is the case, and fixes the few places
in the codebase that were reusing the same function and parameters.

```text
---------------------------------------------------
| name             | met (ms)             | iters |
---------------------------------------------------
| enqueue_directly | 0.01784268040765961  | 65643 |
| compile_first    | 0.017613748867035928 | 67522 |
---------------------------------------------------
```

(0.01784268040765961 - 0.017613748867035928) * 1e6 = 229 nanoseconds

GPU_ORIG_REV_ID: 19359a4ff873ae33875bd5084dcc8efc9e3951ea