[mojo-stdlib] Remove abs/ceil/trunc from FloatLiteral, abs from IntLiteral.
This removes support for the `__abs__` (and ceil and trunc) methods
from float literals, and abs from `IntLiteral`.

We don't have anything that particularly needs the absolute value of
a literal to stay in the infinite precision domain - we can just promote
to Int/Float64.  If we need this in the future, we can add this back,
but until then it is nice to not have the complexity.

MODULAR_ORIG_COMMIT_REV_ID: afc67b82811002583d170a1e0507faf220193e81