[Serving] Remove erroneous finish_reason in streaming responses (#56688)
We currently hardcode a finish_reason in streaming chunk responses. This
is a bug and inconsistent with the open ai spec
(https://platform.openai.com/docs/api-reference/chat/streaming).

It's worth noting that our last chunk is just the string "[DONE]" and
not the expected structured chunk response. This PR does not fix that,
but it at least prevents any clients that look for the finish reason in
a streaming response to detect the end of sequence.

SERVE_ORIG_REV_ID: 4a3e6568490868d5967cafef6d40d0824a822dda