[Driver] Make Accelerator a type in Python (#56738)
Previously `Accelerator` was a factory method that created an object of
type `Device` in Python that was an accelerator in the C++
implementation. This change pushed the `Accelerator` type concept up to
Python, so that `Accelerator` is not a method, but a type now.

This allows type annotations in Python. For example, a list of
accelerators used to be defined like this:

```python
graph_devices: list[DeviceRef]
```

Now it can be defined like this:
```python
graph_devices: list[Accelerator]
```

This allows for a more precise definition in Python.

MAX_PYTHON_ORIG_REV_ID: 230a9e6284a090c3cc60b93f002f6cf04d5e7536