[MOGG][Extensibility] Add _trace_name parameter to the Kernels. (#56266)
Add a new `_trace_name` parameter to the for_ech function (and update all kernels to pass it).
Also add a new `_trace_name` parameter to all kernels using trace.
In the GraphCompiler, we set the parameter to "" first, then update it befoire doing soft fusion.
This breaks HoistHarness because _trace_name parameters all have unique values. This is fixed with a workaround that consist in reseting it to "", then setting it back at the end to its original value.

Thanks to this, now when looking at the traces, we can see the names of the fused ops.

Towards GEX-1872.

MODULAR_ORIG_COMMIT_REV_ID: 7771135bde4fc2ac93e66a0c558b65c35d408c44