[KVCache] rework max_cache_length to max_context_length (#56592)
We were previously passing areound the max number of encoded tokens in
the
cache before QKV projection. This, out, is less useful than the max
context length, which is the sum of the number of encoded tokens in tthe
KVCache plus the new tokens we're encoding in the active forward pass.

This PR cleans this up, and is a prerequisite to work included in
E2EOPT-25t l

KERNEL_BENCHMARKS_ORIG_REV_ID: 2ba4178bd602dabfce4b9fb10181b98e42e241cb