[MOGG] Move kernel slicing logic into MOToKGENLowering (#56631)
The current workflow for kernel slicing is a two-step process where
kernels are sliced in the MOGGPreElab phase. The generated lambdas are
then picked up in MOToKGENLowering if they need to be remapped to new
names. This logic work for operations of a fixed arity, but cannot work
for variadic kernels, as the number of required fusion lambdas is not
known at the MOGGPreElab phase.

This change moves the slicing (generation of input/output fusion
lambdas) into the MOToKGENLowering transform. At that point, the exact
number of arguments for the generated kernel is known.

Towards: GEX-1833
MODULAR_ORIG_COMMIT_REV_ID: c0f8664ec2a08606b014d9ca4910981a1be0f7bf