[mojo-stdlib] Relax the 'atol' and 'rtol' typse of isclose/approx_equal.
This changes the atol/rtol arguments to "close by" checks to take
`Float64` instead of being constrained to having to be the same as
the values being compared.  This constraint adds a bunch of casts
and makes the API more difficult to work with.  All float types
convert to Float64, so this is more flexible.

INTERNAL_UTILS_ORIG_REV_ID: 5604d2c073e916f1fc38755c451bbe5db5971449