[Pipelines] Fix default value for bitmask in null scenario
There was a bug in the default values for the bitmask, in the scenario
in which the server was enabled for structured output, but no json
schema
was passed. This fixes this bug, allowing both structured and
non-structured
queries interspersed.

MODULAR_ORIG_COMMIT_REV_ID: 49514fd65bdb323af1b9df7a0a3e648a053d0fd2