[Support] Fix minor bug in `fillHomegeneous()` (#56448)
### Summary
Fix a probably-not-hit-in-practice bug in `fillHomogeneous()`, which
AFAICT is only used by the model tool's input-generation logic.

### Changes
  - Fix bug
  - Loosen assertion to allow elementsize == buffersize

### Notes

This function has two calls to the `fillFromChunk()` helper. The bottom
one is the common case which knows that we have a full chunk buffer and
can therefore call the function with a static size.

The first call is in the "weird" case where the buffer may or may not be
full. As the comment indicates, this means we have to call the helper
with a dynamic value. In many cases the dynamic value will be equal to
the full buffer size and this change will have no effect, but not
always.

h/t to @zyx-billy who caught this in #56198.

MODULAR_ORIG_COMMIT_REV_ID: ebcf4f52d7b4ee254d9282a0a2941338cefb3254