[KGEN] Prevent CSE of NoneType ParamConstants + Prevent None Captures (#56984)
See MSTDL-1123 for some context.

The motivation behind this is to prevent the capture of "NoneType"
values in KGEN. This edge case has revealed several problems with the
current compilation flow for GPU kernels.

A brief overview of the problem is as below.

In the graph compiler, kernel generation can create kernels with
closures which capture "None". This is normally possible except via CSE
which transforms something that looks like:

```
%none = ...

kgen.param.region.declare = ... {
   %none_again = ...
   ...
   kgen.return %none_again
}

kgen.param.region.declare = ... {
   %none_again = ...
   ...
   kgen.return %none_again
}

// call function which uses the two above closures
```

Into:

```
%none = ...

kgen.param.region.declare = ... {
   ...
   kgen.return %none
}

kgen.param.region.declare = ... {
   ...
   kgen.return %none
}

// call function which uses the two above lambdas
```

This patch prevents this behavior and it appears to be impossible to now
capture "None" for closures.

How does this behavior lead to errors? The error seems to come from a
mismatch between the code which compiles PTX kernels and the code which
outlines and handles the runtime handling of captures. Overall, this
system is a bit hacky as discussed
[Internal link]
however the error in this scenario is slightly different.

Let us look at the signature of the device code we compile and run.
Early in the compilation process of device functions, we might see a
signature like:

`%arg0: !kgen.none, %arg1: ...`

Aka we see the capture for none are in the signature of the function. A
full gist is
[here](https://gist.github.com/AndrewZhaoLuo/c5f4c7ebcdf549a24fc6f99f498e7ed0)

The signature however is dropped when translated to LLVM MLIR dialect:

`llvm.func
@algorithm_functional__elementw6A6A6A6A_6264ac6892919c28b7c294eb87e63c8e(%arg0:
!llvm.struct<(ptr, array<1 x i64>, array<1 x i64>)>`

In the translation process we lower `None` to an [empty LLVM
[Internal link]

Interestingly, if we change this to not return an empty struct, but
maybe something like 8 bytes, this also seems to fix things. By this I
mean if we introspect the PTX, we now see the input parameters match
what we expect:

```
.visible .entry algorithm_functional__elementw6A6A6A6A_6264ac6892919c28b7c294eb87e63c8e(
	.param .align 8 .b8 algorithm_functional__elementw6A6A6A6A_6264ac6892919c28b7c294eb87e63c8e_param_0[8],
	.param .align 8 .b8 algorithm_functional__elementw6A6A6A6A_6264ac6892919c28b7c294eb87e63c8e_param_1[24],
	.param .align 8 .b8 algorithm_functional__elementw6A6A6A6A_6264ac6892919c28b7c294eb87e63c8e_param_2[24],
	.param .align 8 .b8 algorithm_functional__elementw6A6A6A6A_6264ac6892919c28b7c294eb87e63c8e_param_3[24],
	.param .align 8 .b8 algorithm_functional__elementw6A6A6A6A_6264ac6892919c28b7c294eb87e63c8e_param_4[24],
	.param .u64 algorithm_functional__elementw6A6A6A6A_6264ac6892919c28b7c294eb87e63c8e_param_5,
	.param .u64 algorithm_functional__elementw6A6A6A6A_6264ac6892919c28b7c294eb87e63c8e_param_6,
	.param .align 8 .b8 algorithm_functional__elementw6A6A6A6A_6264ac6892919c28b7c294eb87e63c8e_param_7[8],
	.param .u64 algorithm_functional__elementw6A6A6A6A_6264ac6892919c28b7c294eb87e63c8e_param_8
```

Vs without a change:

```
.visible .entry algorithm_functional__elementw6A6A6A6A_1497b4c96d3528f551e274bcfa48ef8c(
	.param .align 8 .b8 algorithm_functional__elementw6A6A6A6A_1497b4c96d3528f551e274bcfa48ef8c_param_0[24],
	.param .align 8 .b8 algorithm_functional__elementw6A6A6A6A_1497b4c96d3528f551e274bcfa48ef8c_param_1[24],
	.param .align 8 .b8 algorithm_functional__elementw6A6A6A6A_1497b4c96d3528f551e274bcfa48ef8c_param_2[24],
	.param .align 8 .b8 algorithm_functional__elementw6A6A6A6A_1497b4c96d3528f551e274bcfa48ef8c_param_3[24],
	.param .u64 algorithm_functional__elementw6A6A6A6A_1497b4c96d3528f551e274bcfa48ef8c_param_4,
	.param .u64 algorithm_functional__elementw6A6A6A6A_1497b4c96d3528f551e274bcfa48ef8c_param_5,
	.param .align 8 .b8 algorithm_functional__elementw6A6A6A6A_1497b4c96d3528f551e274bcfa48ef8c_param_6[8],
	.param .u64 algorithm_functional__elementw6A6A6A6A_1497b4c96d3528f551e274bcfa48ef8c_param_7
```
Where the difference is the first parameter, which is actually present
in the code that works.

Meanwhile, the case without the input parameter is likely wrong. This is
because in the parallel process of outlining kernels for captures, we
assume we are passing in something for the NoneType capture, where in
LLVM it is totally dropped in the signature. This causes the weird
"frameshift" type error we have seen.

Ultimately, there is likely some additional digging at the intended
behavior of PTX / LLVM compilation here; specifically why it essentially
changes the arguments to the ptx kernel.

Alternatives:
- Fix up hacky pipeline which finds and populates captures to account
for this edge case around None
- Dig deeper into figuring out why LLVM modifies the signature of the
function

cc @stumpOS

MODULAR_ORIG_COMMIT_REV_ID: 2654d64ae10c780022d0e1d5b4a018d069653635